requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
gspread>=5.0.0
oauth2client>=4.1.3